feat: add is dangerous key to encode/decode
