+++ src/__tests__/encode-test.js
@@ -1,4 +1,5 @@
 jest.dontMock('../encode');
+jest.dontMock('../isDangerousKey');
 jest.dontMock('../ParseACL');
 jest.dontMock('../ParseFile');
 jest.dontMock('../ParseGeoPoint');
